modify dtype to float16 for variables

+++ DL_scripts/BERT.py
@@ -219,7 +219,7 @@ def create_attention_mask_from_input_mask(from_tensor, to_mask):
     # to_seq_length = tf.shape(from_tensor)[1]
 
     to_mask = tf.cast(
-      tf.reshape(to_mask, [batch_size, 1, to_seq_length]), tf.float32)
+      tf.reshape(to_mask, [batch_size, 1, to_seq_length]), tf.float16)
 
     # We don't assume that `from_tensor` is a mask (although it could be). We
     # don't actually care if we attend *from* padding tokens (only *to* padding)
@@ -227,7 +227,7 @@ def create_attention_mask_from_input_mask(from_tensor, to_mask):
     #
     # `broadcast_ones` = [batch_size, from_seq_length, 1]
     broadcast_ones = tf.ones(
-      shape=[batch_size, from_seq_length, 1], dtype=tf.float32)
+      shape=[batch_size, from_seq_length, 1], dtype=tf.float16)
 
     # Here we broadcast along two dimensions to create the mask.
     mask = broadcast_ones * to_mask
@@ -401,7 +401,7 @@ class AttentionLayer(tf.keras.layers.Layer):
             # Since attention_mask is 1.0 for positions we want to attend and 0.0 for
             # masked positions, this operation will create a tensor which is 0.0 for
             # positions we want to attend and -10000.0 for masked positions.
-            adder = (1.0 - tf.cast(attention_mask, tf.float32)) * -10000.0
+            adder = (1.0 - tf.cast(attention_mask, tf.float16)) * -10000.0
 
             # Since we are adding it to the raw scores before the softmax, this is
             # effectively the same as removing these entirely.
